Move Range, Scale, TicLabels into the axis module.
